// src/app/models/backend.api.model.ts
export enum BackendMethod {
    subscribe = "subscribe",
    editProduct = "editProduct",
    editSmartCollection = "editSmartCollection",
    editCustomCollection = "editCustomCollection"
} 

export enum BackendCollectionState {
    preannouncement = "preannouncement",
    announcement = "announcement",
    presale = "presale",
    onsale = "onsale",
    postsale = "postsale"
}